add new service class

+++ src/main/java/searchengine/dto/statistics/DetailedStatisticsItem.java
@@ -9,6 +9,6 @@ public class DetailedStatisticsItem {
     private String status;
     private long statusTime;
     private String error;
-    private long pages;
-    private long lemmas;
+    private int pages;
+    private int lemmas;
 }
